New program fields : schema > data > typescript > components (#114)
- [x] Rename fields in json schema AND in yaml files
- [x] Update frontend's TS definition for programs
- [x] Update frontend components to display fields' values into html
templates
- [x] Check main tests : 
  - [x] `npm run test-data -w packages/backend` 
  - [x] `npm run build-front`

---------
