+++ packages/web/public/data/programs/diag-ecoconception.yaml
@@ -1,32 +1,29 @@
 id: diag-ecoconception
-title: Diag Ecoconception
-resume: Améliorez l'impact environnementale de vos produits et services
-text: | 
+titre: Diag Ecoconception
+promesse: Améliorez l'impact environnementale de vos produits et services
+description: | 
   Bénéficiez des experts d'un conseil pour repenser la manière de concevoir vos produits, services et procédés.
-origin_url: https://diag.bpifrance.fr/produit/diag-eco-conception
-# cover: images/programs/diag-ecoconception.png
-cover: images/TEE_ampoule.png
-description:
+url: https://diag.bpifrance.fr/produit/diag-eco-conception
+illustration: images/TEE_ampoule.png
+objectifs:
   - Une démarche simple
   - Sensibilisez les équipes internes à l’écoconception et formez un référent écoconception dans votre entreprise
   - Réalisez une évaluation environnementale basée sur l’Analyse du Cycle de Vie du produit, service ou procédé concerné
   - Identifiez, formalisez et chiffrez les leviers d’écoconception, en maîtrisant le risque de transfert de pollution
   - Préparez la mise en œuvre d'un plan d’actions incluant un premier dimensionnement technico-économique
-program_types: 
-  - diag
 conditions:
   - type: project_needs
     operator: "or"
-    value: 
+    value:
       - advices
       - financing
   - type: project_status
     operator: "or"
-    value: 
+    value:
       - improve
   - type: project_sectors
     operator: "or"
-    value: 
+    value:
       - "*"
   - type: structure_sizes
     operator: "or"
@@ -37,8 +34,10 @@ conditions:
       - GE
   - type: user_goals
     operator: "or"
-    value: 
+    value:
       - ecoconception
-program_providers: 
-  - code: BPIFRANCE
-    siren : 320 252 489
+opérateur de contact: Bpifrance
+autres opérateurs:
+  - ADEME
+nature de l'aide: accompagnement
+coût de l'accompagnement: 5 400 € HT après subvention de 70 %
