+++ packages/web/public/data/programs/fonds-tourisme-durable.yaml
@@ -1,44 +1,40 @@
 id: fonds-tourisme-durable
-title: Fonds Tourisme Durable
-resume: Obtenez un diagnostic et des subventions dans votre activité touristique 
-text: | 
-  Le Fonds Tourisme Durable soutient la réalisation d'un diagnostic environnemental gratuit et de financements d'investissements et/ou d'études. 
-origin_url: https://agirpourlatransition.ademe.fr/entreprises/aides-financieres/2023/fonds-tourisme-durable-restaurateurs-hebergeurs-accelerez-transition-0
-# cover: images/programs/fonds-tourisme-durable.jpg
-cover: images/TEE_ampoule.png
-description:
+titre: Fonds Tourisme Durable
+promesse: Obtenez un diagnostic et des subventions dans votre activité touristique
+description: |
+  Le Fonds Tourisme Durable soutient la réalisation d'un diagnostic environnemental gratuit et de financements d'investissements et/ou d'études.
+url: https://agirpourlatransition.ademe.fr/entreprises/aides-financieres/2023/fonds-tourisme-durable-restaurateurs-hebergeurs-accelerez-transition-0
+illustration: images/TEE_ampoule.png
+objectifs:
   - Faites un état des lieux gratuitement afin d'identifier les axes prioritaires d’économies
   - Réduisez et maîtrisez vos coûts fixes (énergie, eau, déchets, gaspillage alimentaire, transport)
   - Accédez à des aides forfaitaires dans tous les domaines de la transition écologique (entre 5k€ et 200k€)
-program_types: 
-  - aap
 conditions:
   - type: project_needs
     operator: "or"
-    value: 
+    value:
       - advices
       - financing
   - type: project_status
     operator: "or"
-    value: 
+    value:
       - "*"
   - type: project_sectors
     operator: "or"
-    value: 
+    value:
       - tourism
   - type: structure_sizes
     operator: "or"
-    value: 
+    value:
       - PME
       - TPE
       - ETI
       - GE
   - type: user_goals
     operator: "or"
-    value: 
+    value:
       - impact
       - consolidation
-program_providers: 
-  - code: ADEME
-    href: https://ademe.fr/
-    SIREN: 385 290 309
+opérateur de contact: ADEME
+nature de l'aide: accompagnement
+coût de l'accompagnement: gratuit
