+++ packages/web/src/components/TeeResults.vue
@@ -40,30 +40,26 @@
     <!-- PROGRAM CARD -->
     <div
       v-for="prog in resultsProgs"
-      :key="prog.index"
+      :key="prog.id"
       class="fr-card fr-enlarge-link fr-card--horizontal-tier fr-mb-10v"
       @click="updateDetailResult(prog.id)">
       <div class="fr-card__body">
         <div class="fr-card__content">
           <h3 class="fr-card__title">
-            {{ prog.resume }} 
+            {{ prog.promesse }} 
           </h3>
-          <!-- <p
-            class="fr-card__desc"
-            v-html="prog.description">
-          </p> -->
           <p
             v-if="debug"
             class="vue-debug fr-card__desc">
             <br> choices.publicPath : <code>{{ choices.publicPath }}</code>
-            <br> prog.cover : <code>{{ prog.cover }}</code>
+            <br> prog.cover : <code>{{ prog.illustration }}</code>
             <!-- {{ `${choices.publicPath}${randomImage()}` }} -->
           </p>
           <div class="fr-card__start">
             <ul class="fr-badges-group">
               <li>
                 <p class="fr-badge fr-badge--info fr-badge--no-icon">
-                  {{ prog.title }}
+                  {{ prog.titre }}
                 </p>
               </li>
             </ul>
@@ -71,27 +67,22 @@
         </div>
       </div>
       <div
-        v-if="prog.cover"
+        v-if="prog.illustration"
         class="fr-card__header">
         <div class="fr-card__img">
           <img 
             class="fr-responsive-img"
-            :src="`${choices.publicPath}${prog.cover}`"
-            :alt="`image / ${prog.title}`"
+            :src="`${choices.publicPath}${prog.illustration}`"
+            :alt="`image / ${prog.titre}`"
             />
           <!-- L’alternative de l’image (attribut alt) doit toujours être présente, 
             sa valeur peut être vide (image n’apportant pas de sens supplémentaire au contexte) 
             ou non (porteuse de texte ou apportant du sens) selon votre contexte -->
         </div>
         <ul class="fr-badges-group">
-          <li
-            v-for="(provider, i) in prog.program_providers"
-            :key="`provider-${i}`"
-            >
-            <p class="fr-badge">
-              {{ provider.code }}
-            </p>
-          </li>
+          <p class="fr-badge">
+            {{ prog['opérateur de contact'] }}
+          </p>
         </ul>
       </div>
     </div>
